+++ extensions/iceworks-project-creator/web/src/components/HeaderTitle/index.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import styles from './index.module.scss';
+
+interface IHeaderTitle {
+  title?: string
+}
+
+const HeaderTitle = ({ title = '' }: IHeaderTitle) => {
+  return (
+    <div className={styles.headerTitle}>
+      {title}
+    </div>
+  );
+};
+
+export default HeaderTitle;