+++ extensions/iceworks-time-master/src/recorders/keystrokeStats/keystrokeStats.ts
@@ -31,7 +31,7 @@ export class KeystrokeStats {
 
   public end: number;
 
-  public files: {[filePath: string]: FileChange} = {};
+  public files: { [filePath: string]: FileChange } = {};
 
   public project: Project;